chore: separate create invoice screen from receive
