+++ app/(app)/receive/invoice.js
@@ -1,0 +1,5 @@
+import { CreateInvoice } from "../../../pages/receive/CreateInvoice";
+
+export default function Page() {
+  return <CreateInvoice />;
+}
