+++ app/(app)/receive/withdraw.js
@@ -1,0 +1,5 @@
+import { Withdraw } from "../../../pages/receive/Withdraw";
+
+export default function Page() {
+  return <Withdraw />;
+}
