--- app/(app)/withdraw/index.js
@@ -1,5 +1,0 @@
-import { Withdraw } from "../../../pages/withdraw/Withdraw";
-
-export default function Page() {
-  return <Withdraw />;
-}
